+++ app/[locale]/(public)/about/content/about_pt.mdx
@@ -2,14 +2,13 @@ import { Button } from '@/components/atoms/button';
 import { Separator } from '@/components/atoms/separator';
 import Image from 'next/image';
 
-import AdvisorsSection from '../components/AdvisorsSection';
-import BeliefsSection from '../components/BeliefsSection';
-import ClinicalExpertsSection from '../components/ClinicalExpertsSection';
-import HeadlineSection from '../components/HeadlineSection';
-import JoinNetworkSection from '../components/JoinNetworkSection';
-import MissionSection from '../components/MissionSection';
-import TeamSection from '../components/TeamSection';
-import TextBlock from '../components/TextBlock';
+import HeadlineSection from '../../../../../components/molecules/HeadlineSection';
+import AdvisorsSection from '../../../../../components/organisms/about/AdvisorsSection';
+import BeliefsSection from '../../../../../components/organisms/about/BeliefsSection';
+import ClinicalExpertsSection from '../../../../../components/organisms/about/ClinicalExpertsSection';
+import JoinNetworkSection from '../../../../../components/organisms/about/JoinNetworkSection';
+import MissionSection from '../../../../../components/organisms/about/MissionSection';
+import TeamSection from '../../../../../components/organisms/about/TeamSection';
 
 <HeadlineSection
   title="Transformando os cuidados de saúde da mulher através da inovação e compaixão"
