+++ components/organisms/about/BeliefsSection.tsx
