+++ components/organisms/about/TeamSection.tsx
@@ -1,9 +1,8 @@
 'use client';
 
 import { Separator } from '@/components/atoms/separator';
+import TextBlock from '@/components/atoms/TextBlock';
 import Image from 'next/image';
-
-import TextBlock from './TextBlock';
 
 type TeamMember = {
   name: string;
